fea: index.html generated by build process
I created a couple more gulp tasks to copy the dev.html into index.html
and to be able to launch the server without live-realoading to simulate
production. The obvious problem here is that now we have to work with
dev.html and index.html gets overwritten on every build, but this is ok
for the exercise and doesn't mess up github pages deploy.